engine streamline and optimize rule indexing; group rules with same rule name in the index to a rule pack; rule condition indexing modulo parameter renaming; parser macros for writing parsers; rewrite the core rule engine parser using new parser macros; data structure for defining algebraic constructor and deconstructor in the rule language; reduce number of function calls, new regions, and region copies for running a rule; fix g++ warnings and errors.
language
distinguish between functional if and logical if; user defined pattern matching; first-order functions; let expression; pattern matching on data in assignment expression; tuples and lifted lifted tuples.

irule 
make irule parse both .ir and .r file depending on the file extension and multiple rules in .r file; make the syntax for multiple lines of rules backward compatible.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@3963 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/server/re/src/msiHelper.c
@@ -739,7 +739,7 @@ msParam_t *stringOut, ruleExecInfo_t *rei)
       strPtr = origStr + origLen + intOffset;
     }
 
-    if (intLength >= 0 && strlen (strPtr) > intLength) {
+    if (intLength >= 0 && ((int) strlen (strPtr)) > intLength) {
     /* put a null at the end of the sub str */
       savedPtr = strPtr + intLength;
       savedChar = *savedPtr;
